refactor(driversenum): move to package
