+++ src/main/java/edu/iastate/ece/sd/sdmay2126/runner/selenium/driver/SeleniumWebDrivers.java
@@ -1,4 +1,4 @@
-package edu.iastate.ece.sd.sdmay2126.runner.selenium;
+package edu.iastate.ece.sd.sdmay2126.runner.selenium.driver;
 
 /**
  * Indicates which Selenium WebDrive and web browser to use.
